+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 3.2.0 - 2025-05-09
+
+### Added
+
+- `Innmind\IO\Files\Temporary::close()`
+- `Innmind\IO\Files\Temporary::push()`
+
 ## 3.1.0 - 2025-04-21
 
 ### Added